Added genetic algorithm with binary representation. (Not finished yet)

+++ Nelder-Mead/nelder_mead.hpp
@@ -6,17 +6,17 @@
 class NelderMead {
 	private:
 	size_t maxIterations;
-	float acc;
-	float real_val;
+	double acc;
+	double real_val;
 
-	float **simplex;
-	float (*function)(const size_t, const float *);
+	double **simplex;
+	double (*function)(const size_t, const double *);
 	size_t N;
 
-	const float r_inc = -0.5f;
-	const float r_exc = 0.5f;
-	const float r_ref = 1.0f;
-	const float r_exp = 2.0f;
+	const double r_inc = -0.5f;
+	const double r_exc = 0.5f;
+	const double r_ref = 1.0f;
+	const double r_exp = 2.0f;
 
 	size_t iterations = 0;
 	int inc_counter = 0;
@@ -26,29 +26,29 @@ class NelderMead {
 	int shr_counter = 0;
 	int reset_counter = 0;
 
-	float *values;
-	float *cm;
-	float *x_inc;
-	float *x_ref;
-	float *x_exc;
-	float *x_exp;
+	double *values;
+	double *cm;
+	double *x_inc;
+	double *x_ref;
+	double *x_exc;
+	double *x_exp;
 
-	float *best_point;
-	float g_minimum;
+	double *best_point;
+	double g_minimum;
 
 	void centerMass();
-	void generatePoint(float *x_op, const float r_op);
-	void shrinkSimplex(float beta);
+	void generatePoint(double *x_op, const double r_op);
+	void shrinkSimplex(double beta);
 
 	int partition(int l, int h);
 	void quicksort(int l, int h);
 
-	void checkBeforeFree(float *other);
+	void checkBeforeFree(double *other);
 	void resetSimplex();
 
 	public:
-	NelderMead(size_t dimensions, float **points, float (*func)(const size_t, const float *));
-	void setParameters(size_t iter, float acc, float real);
+	NelderMead(size_t dimensions, double **points, double (*func)(const size_t, const double *));
+	void setParameters(size_t iter, double acc, double real);
 	~NelderMead();
 
 	void minimize();
